+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# вторник, 01 ноември 2022
+
+- initial right panel ui with dummy data
+
 # вторник, 29 август 2022
 
 - initial skeleton